# Fix selector-id-pattern performance