--- a/.changeset/silly-eagles-grow.md
+++ b/.changeset/silly-eagles-grow.md
@@ -1,0 +1,5 @@
+---
+"stylelint": patch
+---
+
+Fixed: `selector-id-pattern` performance