--- a/lib/rules/selector-id-pattern/index.js
+++ b/lib/rules/selector-id-pattern/index.js
@@ -32,6 +32,8 @@ const rule = (primary) => {
 		const normalizedPattern = isString(primary) ? new RegExp(primary) : primary;
 
 		root.walkRules((ruleNode) => {
+			if (!ruleNode?.selector.includes('#')) return;
+
 			if (!isStandardSyntaxRule(ruleNode)) {
 				return;
 			}