--- a/lib/rules/selector-id-pattern/index.js
+++ b/lib/rules/selector-id-pattern/index.js
@@ -31,9 +31,7 @@ const rule = (primary) => {
 
 		const normalizedPattern = isString(primary) ? new RegExp(primary) : primary;
 
-		root.walkRules((ruleNode) => {
-			if (!ruleNode?.selector.includes('#')) return;
-
+		root.walkRules(/#/, (ruleNode) => {
 			if (!isStandardSyntaxRule(ruleNode)) {
 				return;
 			}